requirements.txt
GitPython>=3.1.40
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
jsbeautifier>=1.14.0
packaging>=23.0
numpy>=1.24.0